ch: support more man pages
Not all man pages are for commands. For example, `git-commit` might not
be a proper command, but `man git-commit` is a real man page.  Simplify
`if`s for checking `cmd_type` and just check exit status of `man` to see
if a page exists or not.  As per `man man`:

    EXIT STATUS
       0      Successful program execution.

       1      Usage, syntax or configuration file error.

       2      Operational error.

       3      A child process returned a non-zero exit status.

       16     At least one of the pages/files/keywords didn't exist or wasn't matched.